Minimal intrusion
Crazy hack to remove some of the interposition: at the end of this the first process is instrumented
and we can record what BIFs are being called, but the scheduler does not get to decide what happens
next.

The next step is to interpose on spawn (to set up things to record), exit and load module things.